requirements.txt
beautifulsoup4==4.12.3
numpy==1.26.4
pandas==2.2.1
plotly==5.19.0
python_dateutil==2.8.2
PyYAML==6.0.1
selenium==4.18.1
streamlit==1.32.2
termcolor==2.4.0
webdriver_manager==4.0.1
